+++ src/slskd/Core/State.cs
@@ -15,7 +15,7 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace slskd
+namespace slskd.Core
 {
     using System;
     using System.Net;
